feat(order): implement crud operation with dto and mapper

+++ src/main/java/com/nak/demo/dto/order/OrderItemResponseDto.java
@@ -12,7 +12,7 @@ public class OrderItemResponseDto {
     private String productName;
 
     @JsonProperty("purchase_amount")
-    private Integer purchaseAmount;
+    private Long purchaseAmount;
 
     @JsonProperty("unit_price")
     private Double unitPrice;
